Crate gerber_types

Gerber commands

This crate implements the basic building blocks of Gerber (RS-274X, aka Extended Gerber version 2) code. It focusses on the low level types and does not do any semantic checking.

For example, you can use an aperture without defining it. This will generate syntactically valid but semantially invalid Gerber code, but this module won't complain.

Minimal required Rust version: 1.6.

Structs

ApertureDefinition

Circle

CoordinateOffset
Coordinate offsets can be used for interpolate operations in circular interpolation mode.

Coordinates
Coordinates are part of an operation.

Polygon

Rectangular

Enums

Aperture

ApertureAttribute

ApertureFunction

Command

CopperType

DCode

Drill

DrillFunction

DrillRouteType

ExtendedCode

ExtendedPosition

FiducialScope

FileAttribute

FileFunction

FilePolarity

FunctionCode

GCode

InterpolationMode

MCode

Operation

Part

Polarity

Position

Profile

QuadrantMode

SmdPadType

StepAndRepeat

Unit

Traits

GerberCode
All types that implement this trait can be converted to Gerber Code.